==> 04_pseudobulk.ipynb <==
# GEX data preprocessing

## 1) Preliminaries

import scanpy as sc
import pandas as pd
import scipy
import os

sc.settings.verbosity = 4

data_folder = '../data'
result_folder = '../Result/PseudoBulk'
graph_folder = '../Graph'

if not os.path.isdir(result_folder):
    os.mkdir(result_folder)

## 2) Read dataset

dataset_name = "complete-gex-dataset-annotated.h5ad"
adata = sc.read_h5ad(os.path.join(data_folder, dataset_name))
adata

adata_raw = adata.raw.to_adata()
adata_raw

adata.obs['PatientTime'] = adata.obs['patient'].astype('str') + '_' + adata.obs['time'].astype('str')
adata_raw.obs['PatientTime'] = adata_raw.obs['patient'].astype('str') + '_' + adata_raw.obs['time'].astype('str')

## 3) Pseudobulk data generation

def adata_pseudobulk(adata):

    if scipy.sparse.issparse(adata.X):
        adata.X=adata.X.todense()

    df_sc = pd.DataFrame(
        index = adata.obs_names.tolist(),
        data = adata.X,
        columns=adata.var_names.tolist()
        )
   
    df_sc = df_sc.sum(axis=0)

    return(df_sc)

total = pd.DataFrame()
obs = pd.DataFrame()

for sample in adata.obs['PatientTime'].unique():
    for cluster in adata.obs['cell_families'].cat.categories:

        dummy = adata[(adata.obs['PatientTime'] == sample) & (adata.obs['cell_families'] == cluster)].copy()
        
        df = adata_pseudobulk(dummy)
        dfmeta=dummy.obs
        total['{}_{}'.format(sample,cluster)] = df
        dfmeta['numberOfCell'] = dummy.shape[0]
        obs = obs.append(dfmeta.loc[:,['PatientTime', 'cell_families', 'patient', 'time', 'numberOfCell', 'severity']].iloc[0,:])

total.to_csv('{}/total.csv'.format(result_folder), sep='\t')
obs.to_csv('{}/obs.csv'.format(result_folder), sep='\t')

total=pd.DataFrame()
obs=pd.DataFrame()

for sample in adata.obs['PatientTime'].unique():
    for cluster in adata.obs['cell_families'].cat.categories:
        
        dummy = adata_raw[(adata_raw.obs['PatientTime'] == sample) & (adata_raw.obs['cell_families'] == cluster)].copy()
        
        df = adata_pseudobulk(dummy)
        dfmeta=dummy.obs
        total['{}_{}'.format(sample,cluster)] = df
        dfmeta['numberOfCell'] = dummy.shape[0]
        obs = obs.append(dfmeta.loc[:,['PatientTime', 'cell_families', 'patient', 'time', 'numberOfCell', 'severity']].iloc[0,:])

total.to_csv('{}/totalRaw.csv'.format(result_folder), sep='\t')
obs.to_csv('{}/obsRaw.csv'.format(result_folder), sep='\t')
